WIP add set-v2p command
Add a commant to set/add a virtual-to-physical mapping in the global
V2P table. Currently limited to IPv4 virtual IP only.

This is actually specific to the Oxide Network, and not a generic
component. However, for the sake of time this is hard-coded in
opteadm/opte-core for the time being. Eventually, commands like this,
as well as command relating to things like firewall, will have some
way to register their APIs/commands with opte-core/opteadm.