Set outer src/dst MAC to zero
The current plan is to have the upcoming xde device perform the IRE
query and fill out the outer src/dst MAC addresses.

In the future this could be moved into OPTE, but it would require a
new hook to run post UFT (as the UFT's job is to cache transformations
for active flows, but the physical routing needs to be dynamic).