replace mac_unicast_add w/ fake mac with MAC_OPEN_FLAGS_NO_UNICAST_ADDR
This was not working before becase I was using MCIS_NO_UNICAST_ADDR on
mac_client_open, which is not a valid flag for mac_client_open, it's the
flag that mac_clietn_open itself uses. Changing this to
MAC_OPEN_FLAGS_NO_UNICAST_ADDR works in the way I was expecting -
setting up a minimal TX path for the mac client that uses the address of
the underlying mac handle.