xde: a mac provider for opte (#48)
This commit adds xde, a mac provider for OPTE. When the xde driver is
loaded, it allows xde devices to be instantiated and configured through
dld ioctls. Xde devices are mac providers, meaning they can be used as
target devices by viona (or vnics for testing), or used directly as
network devices in their own right.  Xde devices process all incoming
and outgoing packets in opte-core according to user-driven configuration
set through ioctls. Each xde device is configured with a pair of
underlay devices. On egress, ire/nce is consulted on what underlay
device to send a packet out, and the destination mac to use to complete
the outer frame for encapsulated packets.

There are some common illumos bits in both the kernel and userspace that
are needed for xde. They are available in the following illumos branch

- https://github.com/oxidecomputer/illumos-gate/tree/xde

As well as a p5p archive ONU build here

- https://github.com/oxidecomputer/os-build/runs/5356524552