XXX opteadm should be a binary only package
Ultimately opteadm should be the binary only, and opte-core-api should
be the API types along with optional features to enable the API
methods. But because opteadm is still reliant on the opte-core types
for some commands we duplicate some logic across projects. The goal is
to move all of the api to opte-core-api. At that point opteadm would
only depend on opte-core-api with the `std` feature enabled (and maybe
some additional features for commands that only make sense in an
operator context, that way we can keep APIs for things like sled-agent
separate from random commands in opteadm).