ext-ip-hack: fwd all ext traffic to phys gateway
This modifies the hacked up NAT layer to always rewrite external
traffic to the local gateway on the "external" subnet. When the
destination is another IP on the subnet the gateway will just forward
it back onto the subnet. In the case of an off-subnet destination, the
gateway will forward it to the next hop. This allows bouth inbound and
outbound external traffic to work with the guest.

I also tried to get intra-guest (aka internal VPC traffic) to work
under the hack but it's fighting me and we want to land what's here
for the next demo.

I also modified the port process DTrace script to print more useful
info. And I added a start to a flowchart for processing flow.