ICMP echo does not work over SNAT (#231)
In order for ICMP Echo to work over SNAT, we must map the Echo
Identifier into the SNAT port space. There are two ways to do this:

1. Map ICMP message bodies into a pseudo-ULP header and allow them to
   be acted upon just like other header data.

2. Add support for body transforms (like VFP's "action context").

OPTE used to implement option (1), but there were some issues in the
past that caused me to remove that behavior. It might be good to
reintroduce that behavior at some point, but instead I opted to take
option (2). We might eventually need body transforms for other
reasons, so I thought it would be good to at least see what it might
look like. This implementation of body transforms is simple: it only
allows transforming of bytes, it doesn't allow a change in body length
(removing or adding bytes). So while this works for the purposes of
implementing ICMP/SNAT rewrite, it may need some further development
for future, more advanced body transformations.

This work only supports outbound ICMP Echo (and associated inbound
reply). There is more ICMP support we need to figure out:

- A DU reply to an ICMP Echo over SNAT. This should be easy enough to
  deal with in the body transform.

- A DU reply to a TCP/UDP flow over SNAT or NAT. A harder problem to
  solves is dealing with DU replies (from some external network) to
  TCP/UDP traffic. The inbound flow id will be for ICMP and therefore
  will match nothing. We could allow all inbound ICMP DU traffic, but
  the user may explicitly block that traffic in the firewall, and we
  need to honor that. Rather, we need to be smart enough to know there
  is a stateful firewall flow and that we need to allow inbound ICMP
  DU for active flows. It could be that we reach into the ICMP DU body
  and use that data to build its flow id, as in some sense **that is**
  its real flow id -- it's a response to **that** flow.

Other Work
----------

- Regression in flow id handling causes it to be immutable (#252)

- Added ability to snapshot (clone) kstats. This is useful for
  verifying stat values in integration tests.