Print port state on test failure
When an integration test fails, it's useful to have the various Port
state printed in a human-friendly manner, similar to how one would
inspect the port with opteadm.

In order to avoid duplicated code we move the print routines into
their respective crates so that they may be shared by both the
integration test as well as opteadm.

I also modified the integration test to create a global VpcMappings,
just like xde currently does. This keeps it closer to how xde
currently works and also reduces the amount of code one needs to write
per test.