Bring router::add_entry into common setup
For most tests, we want the gust to have a route to other guests in
its subnet. Move the router::add_entry() call to the common setup, and
replace routes for the tests that need a different setup.

Added a `chk!` macro. It's like `assert!` but takes the additional
PortAndVps argument so that the port state is printed on failure.

Added `router::replace()` API to allow replacing all routes for a
guest in one, atomic operation.

Removed `check_no_flows()` which was obsoleted by the `VpcPortState`
mechanism.