Add notion of default action
Add the notion of a default action to a layer. In many cases, a layer
is acting as either a filtering layer or a rewrite layer. A filtering
layer provides an allow list of packets that may pass its boundary. A
rewrite layer is meant to rewrite packets that match certain criteria,
but otherwise let the packet pass untouched.

A default action specifies the action to take if a packet matches no
rule in the layer. The default action can be that of allow or deny. Up
to this point we've had an implicit default of allow, requiring
filtering layers to maintain a rule denying any traffic that doesn't
match the allow list. This mechanism for filtering was brittle because
you have to remember to keep the deny rule in place. However, when
allowing a control plane to replace all rules en masse it is easy to
lose this rule by accident. By having the ability to specify a default
action of deny, we can more easily make sure filtering layers (e.g. a
firewall or router) do their job.

This work was spurred by my fear of losing the deny rule for the
outbound side of the routing layer. It felt brittle (and odd) that it
would be up to the control plane to write a rule to drop traffic going
to the any address (`0.0.0.0/0`) if there is no IG for the given
interface, especially since such a rule is not logically part of the
API described by RFD 21. That is, the assumption in a router is that
if there is no match, then there is no route, and the packet must be
dropped. It feels weird to have control plane insert this an drop rule
in the absence of an IG. This could also be done in the API offered
by oxide-vpc, but that also felt just as odd and brittle. Instead, it
felt best to have a default action.

While here I also added a reason value to the layer deny result.