Implement gateway layer
The idea behind the gateway layer is to combine the disparate layers
the currently implement the "virtual gateway". That is, combine arp,
dhcp, icmp, icmpv6, and dhcpv6 into the new gateway layer. This has
the benefit of making the logical purpose of these various rules more
clear (to provide the appearance of a gateway to the guest) as well as
giving us one place to enforce no spoof (as described in RFD 21
§2.8.2. The later is done by setting the default inbound and outbound
actions to "deny" and adding explicit rules to only allow traffic
matching the guest's virtual IP and virtual MAC address.

* Add Action::Allow and Action::StatefulAllow. The former is an
  alternative to Action::Static with the Identity action. The later is
  an alternative to Action::Stateful with the Identity action. They
  make the intention clearer and are a bit more efficient.

* Add a StatefulAllow default action, which is the same as Allow but
  has the added effect of creating Flow Table entries. This is useful
  for the firewall layer where you need an outbound allow to create a
  pair of FT entries in order to implement a "stateful firewall" (yes,
  the word stateful is a bit overloaded here).

* Add default actions to layer dump and list layers commands.

* Fix opte-layer-process.d.

* Create ActionDescEntry as a way to more efficiently and more clearly
  implement a no-op action. This replaces the previous method of using
  the Identity action.

* Modified integration tests to run with the default firewall rules as
  specified by RFD 21 §2.8.1.