Implement UFT pairing
Implement UFT pairing as described in VFP §6.4.1, with the exception
of flow simulation, which isn't implemented yet. This pairing allows
for better UFT invalidation logic and sets us up for truly handling
asymmetric Unified Flow IDs (UFIDs) in the future.

I also wrote a pair of tests to make sure the UFT invalidation logic
works. Though right now there exists a scenario where the outbound UFT
entry will have a `pair` value of `None` even though an inbound UFT
entry exists. This happens when the first packet for a flow arrives
from the inbound direction, as it's only in the inbound direction that
UFT entries can be paired, and we only perform this pairing when
handling the first inbound packet. The reason pairing is only
accomplished on inbound is because of the possibility of asymmetric
UFIDs, as described in VFP §6.4.1. However, this current limitation
does not technically cause any bugs in terms of traffic filtering, as
the epoch mechanism will catch the stale, unpaired outbound entries and
remove them. Once we have "simulation" of flows, this edge case goes
away.

Other work
----------

* Added some convenience functions to UlpMeta.

* Changed TCP_HDR_SZ to TcpHdr::BASE_SIZE.

* Added some test functions for generating HTTP packets.

* Added some of the PortStats to the port_state verification macros.

* Modified incr! to take an array of strings be more consistent with
  update!.

* Make sure to clear action meta between process() calls (in the tests).