Pass ownership of ActionMeta
The fact that `ActionMeta` was passed as a mutable reference to
`Port::process()` was an artifact of some misunderstandings I had
between the C and Rust parts of my brain when doing early work on
opte. There's no reason that ownership shouldn't pass to the process
method. A big benefit of this change is making it much less likely for
leaking metadata between `process()` calls in a test (by forgetting to
call `ActionMeta::clear()`.

You might wonder why ActionMeta is passed as argument at all; why not
just conjure it inside `process()`? Well, it might be useful at some
point to seed the `ActionMeta` with some values before calling into
the port instance. If that turns out not to be true, then allocating
it inside `process()` makes a lot more sense.