Do not panic on zero-byte packet (#197)
My initial thought back in August was to change the `Packet::wrap()`
functions to return a `Result`; and if the length of the packet is
zero bytes, to return an error. However, after spending more time with
this, there's no real issue with wrapping a zero-byte packet. In fact,
it simplifies the code to allow it. Instead, you deal with this
scenario when attempting to parse the packet; returning an error
indicating that the end of the packet has been reached unexpectedly.

While working on this I found that `Packet<Uninitialized>` and
`PacketWriter` weren't really providing much use. Their original
intent was to allocate a new mblk and write the packet from start to
finish, converting it to `Packet<Initialized>` in which state you
could no longer open a `PacketWriter` on it. But the primary use of
such an API is only for testing and packet generation. The typical hot
path doesn't generate packets, but instead wraps and modifies packets
traveling to or from the client. Currently the `Packet::copy()`
function is fine for generating hairpin packets.

Other Work
----------

* Fixed infinite recursion in `EtherHdr` display/debug impl.

* Unified the `Packet::copy()` functions.

* Renamed the `wrap()`/`unwrap()` functions to `wrap_mblk()`/`unwrap_mblk()`
  so that they wouldn't be confused with `Result::unwrap()`.

* Removed the mblk pointer chasing logic. This is more consistent with
  the rest of illumos/mac which performs no such checks.

* Modified `mock_allocb()` to allocate a 16 byte mblk if the requested
  size is 0. This is more in line with how `allocb(9F)` works.

* Added a zero-byte packet test to make sure they are handled correctly.