Handle longer chains of mblk segments
We previously used a statically-sized `heapless::Vec` inside the
`Packet` type. That was used to represent the segments of a larger
`mblk_t`, those linked by the `b_cont` field. This static array had
enough space for 4 elements, and any attempt to push more segments
panicked. We've now hit that.

This uses a straight-up Rust `Vec` instead of the `heapless` variant, to
handle arbitrarily sized linked-lists of `mblk_t`s. This will address
the immediate issue of panics. I've included a regression test for this.

However, we pay the cost of an allocation now. I've taken care that it's
only one, by walking the `mblk_t` chain once to compute the size,
allocating the `Vec`, and then walking the chain again to call
`PacketSeg::wrap_mblk` on each. We may also pay an allocation cost when
rewriting headers in some cases. Both of these should be measured when
we get to performance issues.