prevent zero-sized continuation blocks
When emitting headers, the structure of the message block for the packet
being modified is somtimes altered. If the block holding the current
headers is too small for the new headers, we shrink the old block by the
size of the old headers, allocate a new block for the new headers and
link the new block to the previous block through its continuation
pointer. If the block being shurnk also contains ULP data this is fine.
However, if it's purely a header block, then we wind up shrinking the
old block to zero, and our new block has a zero sized continuation. This
is not a generally expected situation and has caused NIC hardware to
stop working. This commit prevents this from happening.