Avoid panic if there is no external address configured.
We previously panicked if no external IP address was provided at all.
This is because we need a non-zero u32 for creating flow tables, and
unwrap when that size is zero. This commit addresses that by returning
an error instead. We now allow no external IP configuration, in which
case we use a minimum size of 1 for the NAT flow table, though we still
create an empty layer. However, if an IP address is provided, which
technically allows for zero ports (an odd configuration), we return an
error instead of unwrapping.