Only apply 'VNI' metadata to internal packets
Simplest possible solution: we change the meaning of the VNI match
rule to 'traffic originated on this VNI'. The gating factor is that
packets do *not* have the `(0x0129, 0x00)` Geneve TLV appended to them.
This seems to play nicely with the full test suite, but needs the
requisite Dendrite changes.

As an alternative, we could modify sled-agent to push 'VNI = x' and
'Origin = internal', but that would maybe require more far reaching
changes (and a new rule definition).