Fix ICMP flow coalesce behind SNAT
We now solve the above by extracting the ICMP echo ID and treating it as
a pseudo src/dest port for the purposes of UFT matching of ICMP(v6)
flows.

This takes a bunch of extra internal plumbing: ICMP as now treated as a
proper ULP like ICMPv6 is. Accordingly, ICMP ID rewriting on SNAT is now
a header transform rather than a body transform. This does have the
added benefit that we're now making use of the same incremental ULP csum
code, but we do need to take special care around ICMPv4's csum *not*
including the pseudoheader.

Some cleanup still required -- as well as actual tests for another ID
and seeing if we can make the same happen for ICMPv6.

Closes #396.