Fix ICMP flow coalesce behind SNAT (#408)
We now solve this issue by extracting the ICMP echo ID during header parsing
and treating it as a pseudo src/dest port for UFT matching on ICMP(v6) flows.

This takes a bunch of extra internal plumbing: ICMPv4 as now treated as a
proper ULP like ICMPv6 is. Accordingly, ICMP ID rewriting on SNAT is now
a header transform rather than a body transform. This does have the
added benefit that we're now making use of the same incremental ULP csum
code as all other ULPs, but we do need to take special care around ICMPv4's
csum *not* including the pseudoheader.

This includes some amount of cleanup around shared types between ICMP v4
and v6. Use of HTs also makes it easy to add echo handling to `SNat6`, but that's
best left as another PR given the amount of noise that unifying `Snat`/`Snat6` would
add.

Closes #396.