Separate Inbound and Outbound NAT
Inbound NAT has been redesigned so that it will generate an action_desc
bound to the IP which traffic was received on, rather than the
designated ephemeral IP. This will provide the required flow affinity
when there are multiple floating + ephemeral IPs assigned.

Outbound NAT now stores a reference to all available outbound addresses
(on a per-rule basis). In the event that several are available at
the same rule priority, we split them using the CRC of the flow hash.

The main TODO after that is making this dynamic (i.e., adding a new
ioctl for setting network config). Broadly, this should just require
rebuilding the ruleset so shouldn't really need deep-reaching Arcs +
KMutexes.

We are, however, going to run into a tough problem around UFT
invalidation. On any epoch change, currently all UFTs are discarded. The
problem here is that we need to maintain UFTs so that flows don't
reassign themselves to the wrong external IP, even when seemingly
unrelated rulesets (ie., firewall) are changed. There should be some way
to revalidate them against the new layer/rule/action set to preserve
entries where possible.