Initial separation of api::VpcCfg from internal Cfg
There are a few things here in an early form: we want the engine's
definition of VpcCfg to be its own thing rather than relying on
oxide_vpc::api, to simplify storing locks etc. in VpcCfg for more
dynamic state. We also want a shared reference/resource type so that the
central VpcCfg and any built actions can easily update their behaviour
(and any memoized outputs).

While this is *less* necessary for external IP configuration, this work
will better suit changing other aspects to be dynamic such as DHCP, in a
better manner than attempt number 1.

This is very messy and is just enough to get an ioctl working in a clean
way.