Implement first pass at soft flow-table invalidation
There are still some issues I'm aware of, but this meets the criteria
for getting affinity-preservation on external IP add/remove working. The
most obvious is that this *isn't* yet accurately handling most classes
of action which a rule can output, the second is that we perform action
matching effectively twice on a dirty entry to make sure that stats are
working fine.