TCP flow state refactor and fixes (#444)
* Expire TCP Flow table entries over time

* Preliminary TimeWait + SYN escape.

Still a lot to think through. I'm of the opinion that this flow-state
tracking should be best effort and used for early invalidation + stats,
not as a packet filter like it is today.

* Allow TCP flowstate violations

We have two extra mechanisms in play here:
* UnexpectedSegment is now allowed, and logged. This should still
  increment erro counters visible via kstat etc.
* Unexpected segments bearing a SYN will also invalidate the existing
  flow entries and be reprocessed. I'd prefer if we could do this on a
  purely counter-based mechanism, but we'd have to track all layers if
  we wanted to increment the counters for what is definitely a new flow.

This is a move away from TCP traffic filtering/policing. In general,
these mechanisms are covered by the fact that assuming at least one half
is well-behaved, a truly unexpected packet will trigger an RST and
invalidate our state too.

* Need the same logic in `process_x_tcp_new`

* Super WIP, making sure I don't lose state...

* Remove old code, add enums for UFID invariants

* Revisit some decisions.

* Fixes to error tracing

* Nit.

* Better nit. :)

* Test cases.

* Remove unused defs from tests...

* Review feedback and fix inbound NewFlow

Rewriting a comment prompted me to double-check TCP flow state insertion
on inbound packets. It turns out I wasn't adding the TCP flow table
entry after the old one had been removed in such a case -- this is fixed
with an additional test.