Fix Geneve size in `integration_tests::common::encap`
The `encap` method has, for its lifetime, been adding the length of a
UDP header as well as `GeneveHdr::BASE_SIZE` when computing the length
of the Geneve packet metadata for encapsulating test packets. However,
`GeneveHdr` already includes UDP and its fields, so we were requesting
an extra 8 bytes without initialising them. This PR corrects that
length computation.

Fixes #454.