Fix Geneve size in `integration_tests::common::encap` (#456)
* Fix Geneve size in `integration_tests::common::encap`

The `encap` method has, for its lifetime, been adding the length of a
UDP header as well as `GeneveHdr::BASE_SIZE` when computing the length
of the Geneve packet metadata for encapsulating test packets. However,
`GeneveHdr` already includes UDP and its fields, so we were requesting
an extra 8 bytes without initialising them. This PR corrects that
length computation.

Fixes #454.

* Bump rust toolchain as proof of success

* Remove `UdpHdrRaw` from `GeneveHdrRaw`

Following review feedback, this PR removes the raw UDP representation
from the raw Geneve representation. This requires some give and take in
relevant Meta structs to ensure that we're still parsing and emitting
them as a unit.

This does not go so far as splitting UDP out into a separate ULP in
`OuterMeta` for a number of reasons:
 * This would require substantial changes to how EncapPush and other
   actions work to enable (today) a single encap method.
 * Useful encaps are all tied to UDP (Geneve, WireGuard, VXLAN).
 * Most UDP fields can be constructed from the *choice* of encap
   protocol, and we fill in length as the final step regardless so there
   is no use in carrying it on the PacketMeta. (For Geneve and VxLAN at
   any rate).

* Again, `cargo fmt`.

* Minor clippy appeasement