Correctly preserve inbound UFID on outbound TCP close (#469)
* Correctly preserve inbound UFID on outbound TCP close

Outbound TCP packets which transitioned to close were double-removing
the TCP flow entries allocated to them. This is not bad on its own due
to the table lock. However, the inbound UFID used to invalidate both UFT
entries was being sourced from the second remove, so always returned
`None`.

This fix plumbs the inbound UFT down correctly and performs the removal
only once.

Fixes #466.

* Remove TCP UFT entries if no TCP flow entry found

This is an extra safeguard in the event that the TCP flow state's
inbound UFID is never set, and hard-reconciles any case where we have a
UFT for a TCP flow but no TCP flow.

* Minor clippy-driven tweaks.