Fix zone-to-zone (XDE test, kbench local) runner
During setup, we are not doing any work to ensure that Helios has a
valid NDP cache entry ready to use over the simnet link we install for
testing. As a result, XDE selects the right output port, but installs
source and destination MAC addrs of zero.

This worked before; the devices were in promiscuous mode, so the packets
made into `xde_rx`. In other cases, the underlay traffic in e.g. a
SoftNPU deployment was priming all the necessary NCEs, so we always knew
the target MAC address. Obviously this is an easy fix here, and in
practice we'll always have the NCE for the nexthop (i.e., the sidecar).