Fix opte#469 for VPC-external flows
Removing an inbound UFT entry in response to a TCP closure requires
that we remove it using its pre-transform flow ID, rather than
`pkt.flow()`. This is what `TcpFlowEntryState.inbound_ufid` was intended
to capture. To-date, we have been filling in this optional inbound flow
ID on a TCP flow entry using its post-transformation metadata -- either
`pkt.flow()` or `InnerFlowId::from(pmeta)`. This results in a mismatch:
* The UFT is keyed on pre-transform flow ID, e.g., ephemeral IP + remote
  IP.
* The TCP flow table is always keyed on, e.g., private IP + remote IP.

We now correctly use the initial `flow_before` on all inbound TCP flows.