Expose choice of zone brand on kbench remote/server
This is a necessity for testing on newly-built bench gimlet images --
even if sparse is included, I've been unable to get them to come up due
to supposed publisher/incorporation lookup issues.

On the other hand, a helios experiment-image built with extra packages
like iperf will place all those into the baseline where we need them, so
this is the simplest fix to run both halves on sn9/14.