Gateway: allow tx/rx of given CIDRs as required
RFD 21 goes into some detail describing use cases where users might want
to use one instance as a tunnel endpoint for a VPN, and how traffic
towards an arbitrary CIDR can be routed to an instance for, e.g.,
software routing purposes.

To enable this, we need to be able to setup holes in the anti-spoof
filters in the gateway, which are added here as a new pair of ioctls to
add and remove these holes. It's not clear from the RFD whether this is
to be an on/off toggle or as fine-grained as the new ioctls allow, but
we can suoport either case here.