Cache routes received from `next_hop` (#499)
This PR inserts a caching layer in front of `next_hop` to store routes
for a given `(IpAddr6, L4Hash)` for around 100ms, and reorganises
routing-related code into `xde::route`. This is a necessary step as
today we spend around 21% of `xde_mc_tx` in route lookup on a
per-outbound-packet basis. As a consequence, we need install a new
`Periodic` to flush expired entries every 1s, as otherwise we can have a
large amountof detritus pileup from different L4Hash values.

Currently this is implemented as a per-port cache rather than shared
across the driver -- I've done this because it will give us better
sharding of concurrent readers/writers at higher port/guest counts. The
PR as-is pulls us up from ~1.74Gbps to 2Gbps on a single link between
sn9/14.