Correctly scope lock use in `xde_detach` (#545)
As part of #485, we need to *inspect* the contents of the shared
`XdeState` before taking ownership and dropping its contents. However, I
had created a scenario where:

- The lock is acquired, based on a reference derived from the dev info
ptr (`dip`),
- Ownership of the contents of `dip` is taken using `Box::new_raw`,
- These contents (`XdeState`) are freed,
- The `MutexGuard` pointing into the `XdeState` is dropped, and attempts
to unlock a freed mutex.

This PR scopes the `&XdeState` and the derived `MutexGuard` in an inner
block, correcting this drop order issue by forcing the free to happen at
the end of a lexical scope.

Fixes #543.