Remove VNIC workaround (#308)
The OPTE side changes for
https://github.com/oxidecomputer/opte/issues/74 enabling `viona` to use
an xde device directly without an intermediate VNIC. The [viona
side](https://code.illumos.org/c/illumos-gate/+/2558) changes are not
yet merged but wanted to put this up for folks to see.

~~Also includes a fix for
https://github.com/oxidecomputer/opte/issues/70 to enable `snoop` on xde
devices.~~

~~And a slight change to the arp rules for the external ip hack; seems
like after the gateway layering changes incoming ARP packets get dropped
because they hit the firewall layer first. The rules added to explicitly
allow them are in the gateway layer which would only be hit after
getting past firewall. Given all this is a hack anyways, I just modified
the rules to get added to the firewall layer for now.~~

---------

Co-authored-by: Kyle Simpson <kyle@oxide.computer>