Support for omicron internet gateway model (#588)
The Omicron internet gateway model associates IP pools, and subsequently
addresses from those IP pools, with internet gateways. When packets are
directed at an internet gateway through a VPC route, they should take on
a source address from the IP pool that is tied to the target internet
gateway. This trivially happens when there is only one internet gateway
and an instance only has an address from one IP pool. But when there are
multiple IP pools in play, particularly with floating IPs from different
pools, a decision needs to be made about which one to use. That question
is answered by the IP-pool/internet-gateway association in combination
with VPC routes that direct packets at a particular internet gateway.

---------

Co-authored-by: Kyle Simpson <kyle@oxide.computer>