Don't unwrap when removing invalid TCP state from fastpath.
Now that the fastpath temporarily drops (and reacquires) the port lock
when TCP state needs to be invalidated, we opened ourselves up to the
possibility that another packet could have removed this state ahead of
us. Equally, a packet could insert new TCP state which we might
accidentally remove.

This PR removes the unwrap on removal to account for the race, and
only removes TCP flows if they are pointer-equal.

Closes #618.