Review feedback.
Includes some changes discussed out-of-band on Just Including The Atomic
Types in the union definition, getting us away from the explicit
UnsafeCell.

A couple of nice things fell out of this -- we have setup a pattern for
if/when we need the other KStat variants, but we can now deduplicate the
code between test/kernel since we always have an `AtomicU64` now.