Explicitly drop non-echo ICMP at SNAT
Our SNAT layer has limited support for ICMP traffic, and is limited to
using the identifier of Echo Request/Reply traffic as a form of
pseudo-port. For other ICMP traffic, there is no comparable field and
the packet must be dropped.

This PR changes how this is requested, from a `GenDescError` (ordinarily
used for programming errors or resource exhaustion) into a standard
drop. This is more appropriate given that guests are free to (attempt
to) send these other ICMP variants at their leisure, including in
response to an ibound frame on a recently closed port.

Closes #727.