Tunneled LSO/CSO Support (#614)
This PR makes use of in-progress illumos bits to detect common offloads
between underlay devices, and advertises LSO/CSO capabilities on `opte`
ports, relying illumos's emulation functionality to split packets and
insert checksums if not on `cxgbe` (or other compatible NIC).

This allows for a few nice things:
* Guests will send us TCP packets without an L4 checksum and up to
~64KiB in size.
   - The guest and OPTE spend less time computing/updating checksums.
- OPTE has to process fewer packets – hooray! Ultimately, that's less
processing time per payload byte.
- If sending out over the NIC, then either the NIC or illumos's LSO
emulation path are responsible for splitting the packet into TCP packets
which will not violate the MTU. See below on what we tweak here.
- If sending in the loopback path, we get to hand these packets directly
to the target guest without splitting them apart. We do need to insert
the checksum ourself, however.
 * Guests will send us UDP packets without an L4 checksum.
- Again, the guest and OPTE spend less time computing/updating
checksums.

Ultimately, we have control over the MSS we advertise to the NIC for
LSO. The useful part of this is that when we know we are using a purely
rack-internal path, we can elevate the MSS up to `MTU - overheads`. I've
added a system of 'well-known' `ActionMeta` KV pairs which allow for
layers like `overlay` to propagate this knowledge out. Given that the
use of a larger MTU vastly reduces inbound packet rate on the receive
half (the main bottleneck today), this gets us to around 8Gbps iPerf for
rack-local traffic (and 17Gbps for two or more parallel streams) on
`glasgow`.

On `dublin` with a full control plane, this resolves to around 4Gbps
sled-to-sled and 14-Gbps sled-local between Linux VMs – we want to
investigate the drop here once this is on dogfood. illumos as a guest
doesn't do much better than before since it does not advertise GRO
(illumos-as-host will fragment the packets) or LSO (it will never send
TCP packets > MTU).

- [x] Cleanup.
- [x] Finalise illumos interfaces.
[stlouis#663](https://code.oxide.computer/c/illumos-gate/+/485)
- [x] Manually fill checksums in guest loopback cases.
- [x] Have viona perform LSO if guests do not advertise/expect LRO.
[illumos#17032](https://code.illumos.org/c/illumos-gate/+/3942)

Closes #328, closes #329.

Based on #688.