Apply NAT to headers/packets carried within ICMP messages (#736)
This PR modifies nested packets in ICMP payloads to ensure that private
source/destination addresses are replaced with the correct external IP.
This makes use of a body transform to modify the packet contents, which
has the downside of downgrading NAT'd ICMP packets to the
less-fast-path. Note that this is not performance-critical traffic, and
these frames are still taking a read lock over the port at worst.

There's a bit of trouble passing the full header stack back into these
body transforms -- we have an arbitrary `T` which can be treated as a
byteslice, but our `dyn BodyTransform`s presumably need to be
monomorphised. We need to hand at least some knowledge of the parsed
headers over, such that we only attempt this transform on the right ICMP
types. For now, we achieve this by passing in a reference to the owned
version of the inner ULP header.

Fixes #369.