First stab at InnerFlowId changes
One of the side effects of wanting to filter on ICMP type is that flow
IDs need to match that level of specificity. With the old FlowId design,
we get *ICMP* and that's it, with the inner ID of an echo paylad added
in occasionally. This can lead to fun cases like an allowed ICMP DU
opening the door for *all* ICMP packets from that remote host.

What we're doing now is providing access to the same two `u16`s in
different contexts via `l4_info`/`_mut`, which lets us store the
type/code in the source port field (cast as two `u8`s). This does need
us to formalise some things we'd taken for granted, e.g., the `mirror`
flow for an ICMP echo/request reply needs to also change out the
expected type.