Feedback (fin.)
I've come down on the side of 'these are fatal programming errors, from
which we cannot/should not attempt to recover'. While I toyed with the
idea of checking strong/weak reference counts ahead of time, it's
brittle in the event that we do want to use `Weak`s to these pointer
types at any point in future (and also inherently *very race condition
friendly*).