Use `Token`-based locking on XDE management operations (#761)
This PR introduces a `TokenLock` type, which allows for a single thread
to be in a critical section without actively holding a `KMutex` or a
write on a `KRwLock`. This is used to ensure that we have at most one
active ioctl handler performing management operations. Previously, we
relied upon write locks to the `underlay` and `xde_devs` to enforce this
constraint.

Underlay installation, port creation, and port destruction have been
restructured to rely on the `TokenLock` as the main means of mutual
exclusion, and then to access other locked resources (which may be
shared with the datapath) more granularly. This allows us to ensure that
any calls to DLS (link name resolution, device creation) are made
without inappropriately holding any lock, but are still appropriately
atomic with respect to one another.

Fixes #758.